pyproject.toml
[tool.poetry]
name = "scripts"
version = "0.1.0"
description = ""
authors = ["Andre Luiz Monteiro da Rocha <andrecorumba@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
llama-index = "^0.9.41"
python-dotenv = "^1.0.1"
setuptools = "^69.0.3"
pypdf = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
